bayes-backend/pyproject.toml
[project]
name = "bayes-backend"
version = "0.1.0"
description = "Bayes AI Tutoring Assistant Backend"
requires-python = ">=3.9"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn==0.32.1",
    "python-dotenv==1.0.1",
    "google-generativeai==0.8.3",
    "pydantic==2.10.4",
    "python-multipart==0.0.20",
    "httpx>=0.26,<0.28",
    "supabase==2.10.0",
    "python-jose[cryptography]==3.3.0",
    "sympy==1.12",
    "stripe>=12.3.0",
    "Pillow>=10.0.0",
    "resend==2.4.0",
    "pdf2image==1.17.0",
    "sse-starlette==2.4.1",
] 
